requirements.txt
blinker==1.9.0
click==8.2.0
Flask==2.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psutil==7.0.0
python-pam==2.0.2
six==1.17.0
Werkzeug==3.1.3
